Fix overflow problem

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -1,5 +1,4 @@
 #include "holberton.h"
-#include <stdio.h>
 /**
  *mapint - maps a character to its integer form
  *@a: character to be mapped
@@ -29,7 +28,7 @@ int isDigit(char i)
  */
 int _atoi(char *s)
 {
-	int num = 0, sign = 1, started = 0;
+	signed long num = 0, sign = 1, started = 0;
 
 	while (*s)
 	{
